Output folder cleanup fix

+++ allure-csharp-commons-v2/Allure.Configuration.Tests/ConfigurationTests.cs
@@ -1,4 +1,6 @@
 ﻿using Allure.Commons;
+using Microsoft.Extensions.Configuration;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -8,62 +10,72 @@ using System.Threading.Tasks;
 using Xunit;
 namespace Allure.Configuration.Tests
 {
-    public class ConfigurationTests : IDisposable
+    public class ConfigurationTests
     {
-        public ConfigurationTests()
+        [Theory(DisplayName = "Should initialize using configuration defaults")]
+        [InlineData(null)]
+        [InlineData(@"{}")]
+
+        public void Defaults(string config)
         {
-            RestoreConfig();
-        }
-        public void Dispose()
-        {
-            RestoreConfig();
-            GC.SuppressFinalize(this);
+            RestoreState(config);
+
+            var allureCycle = new AllureLifecycle();
+
+            Assert.Equal(Path.Combine(Environment.CurrentDirectory, AllureConstants.DEFAULT_RESULTS_FOLDER), allureCycle.Output);
+
         }
 
-        [Fact(DisplayName = "Should create 'Allure-results' folder without config file")]
-        public void ShouldCreateResultsFoldernWithoutConfigFile()
+        private static void RestoreState(string config)
         {
-            if (File.Exists(AllureConstants.CONFIG_FILENAME))
-                File.Delete(AllureConstants.CONFIG_FILENAME);
-
             if (Directory.Exists(AllureConstants.DEFAULT_RESULTS_FOLDER))
                 Directory.Delete(AllureConstants.DEFAULT_RESULTS_FOLDER, true);
 
-            var allureCycle = new AllureLifecycle();
-
-            Assert.True(Directory.Exists(AllureConstants.DEFAULT_RESULTS_FOLDER));
+            if (config != null)
+                File.WriteAllText(AllureConstants.CONFIG_FILENAME, config);
+            else
+                if (File.Exists(AllureConstants.CONFIG_FILENAME))
+                File.Delete(AllureConstants.CONFIG_FILENAME);
         }
 
-        [Fact(DisplayName = "Shouldn't cleanup 'Allure-results' by default")]
-        public void ShouldntCleanupResultsFolder()
+        [Theory(DisplayName = "Should cleanup temp results folder")]
+        [InlineData(@"{""allure"":{""directory"": ""c:\\windows\\allure-results"", ""cleanup"":  true}}")]
+        public void ShouldCleanupTempResultsFolder(string config)
         {
-            if (File.Exists(AllureConstants.CONFIG_FILENAME))
-                File.Delete(AllureConstants.CONFIG_FILENAME);
+            RestoreState(config);
 
-            if (!Directory.Exists(AllureConstants.DEFAULT_RESULTS_FOLDER))
-                Directory.CreateDirectory(AllureConstants.DEFAULT_RESULTS_FOLDER);
-
-            File.WriteAllText(
-                Path.Combine(AllureConstants.DEFAULT_RESULTS_FOLDER,
-                Guid.NewGuid().ToString()),
-                "");
+            var expectedResultsFolder = Path.Combine(Path.GetTempPath(), AllureConstants.DEFAULT_RESULTS_FOLDER);
+            File.WriteAllText(Path.Combine(expectedResultsFolder, Guid.NewGuid().ToString()), "");
 
             var allureCycle = new AllureLifecycle();
-            Assert.NotEmpty(
-                Directory.GetFiles(AllureConstants.DEFAULT_RESULTS_FOLDER));
+
+            Assert.Empty(Directory.GetFiles(expectedResultsFolder));
+        }
+
+        [Theory(DisplayName = "Should cleanup existing results folder")]
+        [InlineData(@"{""allure"":{""cleanup"": true}}")]
+
+        public void ShouldCleanupExistingResultsFolder(string config)
+        {
+            RestoreState(config);
+
+            var expectedResultsFolder = AllureConstants.DEFAULT_RESULTS_FOLDER;
+            Directory.CreateDirectory(expectedResultsFolder);
+            File.WriteAllText(Path.Combine(expectedResultsFolder, Guid.NewGuid().ToString()), "");
+
+            var allureCycle = new AllureLifecycle();
+
+            Assert.Empty(Directory.GetFiles(expectedResultsFolder));
         }
 
         [Fact(DisplayName = "Should access Configuration")]
         public void ShouldAccessConfigProperties()
         {
+            var config = @"{""allure"":{""customKey"": ""customValue""}}";
+            RestoreState(config);
             var cycle = new AllureLifecycle();
             Assert.Equal("customValue", cycle.Configuration["allure:customKey"]);
 
-        }
-
-        private static void RestoreConfig()
-        {
-            File.Copy("testConfig.json", AllureConstants.CONFIG_FILENAME, true);
         }
     }
 }
